Utilize webpack-dev-middleware for hosting
The prior solution of pushing all content to the to the webpack server broke down for globs and watching. Karma really desires files and also prefers one to one mapping of said files and trying to fight that was not fruitful.

Instead opted to move all entry points back to the normal Karma serving routes and any auxiliary resources through the middleware.

This also removes much of the complexity and potential overhead relating to tracking individual file changes as Karma does not seem to do incremental testing and a filelist refresh will ensure that the rebuild occurs.

Fixes #1