feat: (watch) Allow watchmode to rerun only tests in file changed
Currently a file change results in all tests being rerun. This change allows only the tests in the changed file to be rerun. If the changed file is not a test file then all tests will rerun.

This change is a copy of the work done in this fork https://github.com/aaronjensen/karma-webpack. However it is adapted to the current version of karma