fix: honor webpack's `bail` by emulation inside plugin
As long as #66 (which is about `bail` not being honored by webpack) is
not fixed, we can emulate equivalent behavior inside the plugin as an
alternative solution. If the webpack compilation step finishes with
errors and the `bail` option is set to true, we exit Karma with a
nonzero exit status. This way, we can still prevent Karma from silently
continuing upon webpack compilation errors.