feat(ci): added github workflow for testing
this will be run against pull requests on next
and master.

this also updates karma to v6 and utilizes the new
public api so we no longer have to fork a child
process to avoid karma < 6's behavior of calling
process.exit(#)

this also increases the default timeout of the
integration tests as 5 seconds was just too short
for the CI to be able to properly complete.

Fixes N/A